+++ nengo/__init__.py
@@ -60,7 +60,6 @@ from .rc import rc, RC_DEFAULTS
 from .simulator import Simulator
 from .synapses import Alpha, LinearFilter, Lowpass, Triangle
 from .transforms import Convolution, Dense, Sparse
-from .utils.logging import log
 from . import dists, exceptions, networks, presets, processes, spa, utils
 
 logger = logging.getLogger(__name__)