+++ nengo/tests/test_simulator.py
@@ -1,3 +1,4 @@
+import logging
 import pkg_resources
 
 import numpy as np
@@ -191,7 +192,7 @@ def test_signal_init_values(RefSimulator):
         assert np.all(np.array([1, 2, 3]) == sim.signals[array])
 
 
-def test_seeding(RefSimulator, logger, allclose):
+def test_seeding(RefSimulator, allclose):
     """Test that setting the model seed fixes everything"""
 
     #  TODO: this really just checks random parameters in ensembles.
@@ -219,8 +220,8 @@ def test_seeding(RefSimulator, logger, allclose):
         for attr in attrs:
             check = allclose(getattr(obj1, attr), getattr(obj2, attr)) == equal
             if not check:
-                logger.info("%s: %s", attr, getattr(obj1, attr))
-                logger.info("%s: %s", attr, getattr(obj2, attr))
+                logging.info("%s: %s", attr, getattr(obj1, attr))
+                logging.info("%s: %s", attr, getattr(obj2, attr))
             assert check
 
     ens_attrs = BuiltEnsemble._fields